+++ src/__tests__/types.test-d.ts
@@ -1,0 +1,19 @@
+import { create, ResourceHandle } from '../';
+
+describe('Utility types', () => {
+  describe('ResourceHandle', () => {
+    it('infers the correct type', async () => {
+      async function Test() {
+        return {
+          value: { hello: 'world' },
+        };
+      }
+
+      const test = await create(Test);
+
+      expectTypeOf(test).toEqualTypeOf<ResourceHandle<typeof Test>>({
+        hello: 'world',
+      });
+    });
+  });
+});