+++ CHANGELOG.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - If a resource fails while being destroyed, now its child resources are destroyed as well.
 - Resources can no longer provision child resources after teardown. This closes a loophole where resources could escape destruction.
 
+### Added
+
+- New `ResourceHandle<T>` utility type represents the value returned when creating a resource.
+
 ## [0.4.1] - 2023-01-14
 
 ### Fixed